PRSD-943: Moves PropertyComplianceJourneyDataExtensions into journeyDataExtensions folder
